lp::math
